requirements.txt
accelerate==0.26.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.41.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
datasets==2.16.1
deepspeed==0.13.1
dill==0.3.7
diskcache==5.6.3
exceptiongroup==1.2.0
fastapi==0.109.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
h11==0.14.0
hjson==3.1.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
lit==17.0.6
llama_cpp_python==0.2.43
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.2
pandas==2.0.3
peft==0.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pydantic==1.10.11
pydantic_core==2.16.2
pynvml==11.5.0
pypandoc_binary==1.13
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.1.99
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
sse-starlette==2.0.0
starlette==0.36.3
starlette-context==0.3.6
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.14.1
torch==2.0.1
tqdm==4.66.1
transformers==4.35.0
triton==2.0.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
xxhash==3.4.1
yarl==1.9.4
